creating models files
